+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# 2.1.0
+# 2.1.3
 
-- Added support for custom format link.
-- Removed `author` filter on item to include other type of creator.
+- Added a variable for <authorslastnames> which is a comma delineated string of author's last name
+- Added automatic creation of citations formatted in accordance to citation style from Zotero Citation Repository--Zotero >= 7 only
+- Added a plugin setting that accepts the name of a format form the Zotero Citation Repository (e.g. apa, chicago-author-date) 
+- Added a plugin setting to allow the customization of the autocomplete hint--Zotero >= 7 only
+- Added a plugin setting to allow the customization of the autocomplete hint sort order--Zotero >= 7 only
+- Added a plugin setting to allow the customization of the addition details to show with the auto complete hint--Zotero >= 7 only